+++ core/dbt/exceptions.py
@@ -628,6 +628,20 @@ def macro_not_found(model, target_macro_id):
         .format(model.unique_id, target_macro_id))
 
 
+def macro_invalid_dispatch_arg(macro_name) -> NoReturn:
+    msg = '''\
+    The "packages" argument of adapter.dispatch() has been deprecated.
+    Use the "macro_namespace" argument instead.
+
+    Raised during dispatch for: {}
+
+    For more information, see:
+
+    https://docs.getdbt.com/reference/dbt-jinja-functions/dispatch
+    '''
+    raise_compiler_error(msg.format(macro_name))
+
+
 def materialization_not_available(model, adapter_type):
     materialization = model.get_materialization()