+++ test/integration/008_schema_tests_test/local_utils/macros/current_timestamp.sql
@@ -1,5 +1,5 @@
 {% macro current_timestamp() -%}
-  {{ return(adapter.dispatch('current_timestamp', packages = local_utils._get_utils_namespaces())()) }}
+  {{ return(adapter.dispatch('current_timestamp')) }}
 {%- endmacro %}
 
 {% macro default__current_timestamp() -%}