+++ test/integration/008_schema_tests_test/local_utils/macros/datediff.sql
@@ -19,7 +19,7 @@
     {% if datepart == 'year' %}
         (date_part('year', ({{second_date}})::date) - date_part('year', ({{first_date}})::date))
     {% elif datepart == 'quarter' %}
-        ({{ local_utils.datediff(first_date, second_date, 'year') }} * 4 + date_part('quarter', ({{second_date}})::date) - date_part('quarter', ({{first_date}})::date))
+        ({{ adapter.dispatch('datediff', 'local_utils')(first_date, second_date, 'year') }} * 4 + date_part('quarter', ({{second_date}})::date) - date_part('quarter', ({{first_date}})::date))
     {% else %}
         ( 1000 )
     {% endif %}