+++ test/integration/008_schema_tests_test/test_utils/macros/custom_test.sql
@@ -1,7 +1,7 @@
-{% macro test_pkg_and_dispatch(model) -%}
-  {{ return(adapter.dispatch('test_pkg_and_dispatch', macro_namespace = 'test_utils')()) }}
+{% macro test_dispatch(model) -%}
+  {{ return(adapter.dispatch('test_dispatch', macro_namespace = 'test_utils')()) }}
 {%- endmacro %}
 
-{% macro default__test_pkg_and_dispatch(model) %}
-    select {{ test_utils.current_timestamp() }}
+{% macro default__test_dispatch(model) %}
+    select {{ adapter.dispatch('current_timestamp', macro_namespace = 'test_utils')() }}
 {% endmacro %}