+++ test/integration/012_deprecation_tests/test_deprecations.py
@@ -35,44 +35,6 @@ class TestDeprecations(BaseTestDeprecations):
         self.assertEqual(expected, deprecations.active_deprecations)
 
 
-class TestDispatchPackagesDeprecation(BaseTestDeprecations):
-    @property
-    def models(self):
-        return self.dir('dispatch-models')
-
-    @property
-    def project_config(self):
-        return {
-            'config-version': 2,
-            "macro-paths": [self.dir('dispatch-macros')],
-            "models": {
-                "test": {
-                    "alias_in_project": {
-                        "alias": 'project_alias',
-                    },
-                    "alias_in_project_with_override": {
-                        "alias": 'project_alias',
-                    },
-                }
-            }
-        }
-
-    @use_profile('postgres')
-    def test_postgres_adapter_macro(self):
-        self.assertEqual(deprecations.active_deprecations, set())
-        self.run_dbt()
-        expected = {'dispatch-packages'}
-        self.assertEqual(expected, deprecations.active_deprecations)
-
-    @use_profile('postgres')
-    def test_postgres_adapter_macro_fail(self):
-        self.assertEqual(deprecations.active_deprecations, set())
-        with self.assertRaises(dbt.exceptions.CompilationException) as exc:
-            self.run_dbt(['--warn-error', 'run'])
-        exc_str = ' '.join(str(exc.exception).split())  # flatten all whitespace
-        assert 'Raised during dispatch for: string_literal' in exc_str
-
-
 class TestPackageRedirectDeprecation(BaseTestDeprecations):
     @property
     def models(self):