+++ test/unit/test_macro_calls.py
@@ -24,11 +24,6 @@ class MacroCalls(unittest.TestCase):
                     from {{ model }} )
                 {% endmacro %}""",
             "{% macro test_my_test(model) %} select {{ dbt_utils.current_timestamp() }} {% endmacro %}",
-            "{% macro test_pkg_and_dispatch(model) -%} {{ return(adapter.dispatch('test_pkg_and_dispatch1', packages = local_utils._get_utils_namespaces())()) }} {%- endmacro %}",
-            "{% macro some_test(model) -%} {{ return(adapter.dispatch('test_some_kind1', ['foo_utils1'])) }} {%- endmacro %}",
-            "{% macro some_test(model) -%} {{ return(adapter.dispatch('test_some_kind2', packages = ['foo_utils2'])) }} {%- endmacro %}",
-            "{% macro some_test(model) -%} {{ return(adapter.dispatch(macro_name = 'test_some_kind3', packages = ['foo_utils3'])) }} {%- endmacro %}",
-            "{% macro test_pkg_and_dispatch(model) -%} {{ return(adapter.dispatch('test_pkg_and_dispatch2', packages = (var('local_utils_dispatch_list', []) + ['local_utils2'])) (**kwargs) ) }} {%- endmacro %}",
             "{% macro some_test(model) -%} {{ return(adapter.dispatch('test_some_kind4', 'foo_utils4')) }} {%- endmacro %}",
             "{% macro some_test(model) -%} {{ return(adapter.dispatch('test_some_kind5', macro_namespace = 'foo_utils5')) }} {%- endmacro %}",
         ] 
@@ -40,11 +35,6 @@ class MacroCalls(unittest.TestCase):
             ['get_columns_in_query'],
             ['get_snapshot_unique_id'],
             ['dbt_utils.current_timestamp'],
-            ['test_pkg_and_dispatch1', 'foo_utils4.test_pkg_and_dispatch1', 'local_utils.test_pkg_and_dispatch1', 'local_utils._get_utils_namespaces'],
-            ['test_some_kind1', 'foo_utils1.test_some_kind1'],
-            ['test_some_kind2', 'foo_utils2.test_some_kind2'],
-            ['test_some_kind3', 'foo_utils3.test_some_kind3'],
-            ['test_pkg_and_dispatch2', 'foo_utils4.test_pkg_and_dispatch2', 'local_utils2.test_pkg_and_dispatch2'],
             ['test_some_kind4', 'foo_utils4.test_some_kind4'],
             ['test_some_kind5', 'foo_utils5.test_some_kind5'],
         ]